restructured lib folder - left only initializer in lib dir - moved all internal libs into lib/base folder It should fix issue with loading file 'support.rb'. The description of the problem can be found here: http://gilesbowkett.blogspot.com/2009/04/unshiftfiledirnamefile.html
