+++ lib/base/net_fix.rb
