+++ lib/base/support.rb
