+++ lib/base/version.rb
